Fix async actions for deal containers
